tests/requirements.txt

# BestStag v9.1 - Dependências de Teste
# Pacotes necessários para execução da suite de testes

# Framework de testes principal
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-html>=3.2.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0  # Para execução paralela
pytest-rerunfailures>=12.0  # Para retry de testes flakey

# Testes de performance e carga
locust>=2.16.0
pytest-benchmark>=4.0.0

# Mocks e fixtures
responses>=0.23.0
httpx>=0.24.0
aioresponses>=0.7.0

# Validação e análise
bandit>=1.7.0  # Análise de segurança
safety>=2.3.0  # Verificação de vulnerabilidades

# Relatórios e métricas
coverage>=7.2.0
pytest-json-report>=1.5.0

# Utilitários para testes
faker>=19.0.0  # Geração de dados fake
factory-boy>=3.3.0  # Factory pattern para testes
freezegun>=1.2.0  # Mock de tempo/data

# Testes de API
requests-mock>=1.11.0

# Análise de código
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Dependências específicas do projeto
redis>=4.6.0
psycopg2-binary>=2.9.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Para testes de integração com Docker
docker>=6.1.0
testcontainers>=3.7.0

# Utilitários adicionais
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
